Directory exists check added and create directories accordingly.

+++ data_preprocessing/convert_to_nii.py
@@ -1,3 +1,8 @@
+'''
+OASIS 1 files are stored as dual format .hdr and .img
+Hence this scripts converts the images into .nii format.
+'''
+
 import os
 import nibabel as nib
 import pandas as pd
@@ -17,10 +22,14 @@ for i in range(1,13):
                         print(f'basepath={basep}')
                         img_name=basep+'/RAW/'+p+'_mpr-1_anon.img'
                         print(f'filename={img_name}')
-                        #im = nib.load(img_name)
+			
+                        im = nib.load(img_name)
                         dest=img_name.replace('.img', '.nii')
-                        #nib.save(im,dest)
-                        combined_path_dir=combined_path+p
-                        #os.mkdir('mkdir '+combined_path_dir)
+                        nib.save(im,dest)
+                        
+			combined_path_dir=combined_path+p
+			
+			if not os.path.exists(combined_path_dir):
+                        	os.mkdir(combined_path_dir)
                         print(f"destination={os.path.join(combined_path_dir,dest.split('/')[-1])}")
                         os.system('cp '+dest+' '+os.path.join(combined_path_dir,dest.split('/')[-1]))
